- Removed connection checkin and checkout and replaced it by simply getting the connection from the model. This is the first solution mentioned in the rails connection pool manual (http://api.rubyonrails.org/classes/ActiveRecord/ConnectionAdapters/ConnectionPool.html) instead of the second one.
The problem with the second solution is that when the transaction manager yields the block supplied to it (containing ruby code that modifies the database)
it will use another connection object then the one that is checked in and used for the XA database statements. When the XA statements are executed on another
connection than the normal db statements (insert/update etc) a rollback is pointless and will not work.

The first solution passes the tests successfully but some sources suggest it is tricky. See comment by 'leente' at
http://apidock.com/rails/ActiveRecord/Base/connection

Apparantly, connections can be checked in automatically and thus can be used by another thread.